+++ apiserver/common/src/main/java/com/zipline/global/exception/region/RegionException.java
@@ -1,7 +1,7 @@
-package com.zipline.global.exception.custom.region;
+package com.zipline.global.exception.region;
 
 import com.zipline.global.exception.BaseException;
-import com.zipline.global.exception.custom.region.errorcode.RegionErrorCode;
+import com.zipline.global.exception.region.errorcode.RegionErrorCode;
 
 public class RegionException extends BaseException {
     public RegionException(RegionErrorCode errorCode) {
